Extract checkout cart buttons into a separate component - CheckoutCartButtons
