+++ __tests__/components/checkout/checkout-cart-total.spec.js
@@ -8,7 +8,6 @@ test("renders the cart's total, subtotal and shipping cost", () => {
   // Act
   const wrapper = mount(
     <CheckoutCartTotal
-      continueButtonProps={{}}
       discountSummaries={[]}
       shippingTotal={5}
       subTotal={15}
@@ -27,7 +26,6 @@ test('renders promotions', () => {
   // Act
   const wrapper = mount(
     <CheckoutCartTotal
-      continueButtonProps={{}}
       discountSummaries={[{
         id: 1,
         name: 'Christmas Special',
@@ -47,7 +45,6 @@ test('renders shipping promotions', () => {
   // Act
   const wrapper = mount(
     <CheckoutCartTotal
-      continueButtonProps={{}}
       discountSummaries={[]}
       shippingDiscount={-5}
       shippingDiscountName='Free shipping'
@@ -65,7 +62,6 @@ test('renders payment errors', () => {
   // Act
   const wrapper = mount(
     <CheckoutCartTotal
-      continueButtonProps={{}}
       discountSummaries={[]}
       paymentError='Payment authentication failed'
       shippingTotal={5}
@@ -77,44 +73,4 @@ test('renders payment errors', () => {
   // Assert
   expect(wrapper).toMatchSnapshot()
   expect(wrapper).toIncludeText('Payment authentication failed')
-})
-
-test('displays a correct disabled button based on passed props', () => {
-  // Act
-  const wrapper = mount(
-    <CheckoutCartTotal
-      continueButtonProps={{
-        label: 'Choose Shipping Method',
-        disabled: true
-      }}
-      discountSummaries={[]}
-      shippingTotal={5}
-      subTotal={15}
-      total={20}
-    />
-  )
-
-  // Assert
-  expect(wrapper.find('button.c-cart-summary-buttons__cta--proceed')).toHaveText('Choose Shipping Method')
-  expect(wrapper.find('button.c-cart-summary-buttons__cta--proceed')).toHaveProp('disabled', true)
-})
-
-test('displays a correct enabled button based on passed props', () => {
-  // Act
-  const wrapper = mount(
-    <CheckoutCartTotal
-      continueButtonProps={{
-        label: 'Choose Shipping Method',
-        disabled: false
-      }}
-      discountSummaries={[]}
-      shippingTotal={5}
-      subTotal={15}
-      total={20}
-    />
-  )
-
-  // Assert
-  expect(wrapper.find('button.c-cart-summary-buttons__cta--proceed')).toHaveText('Choose Shipping Method')
-  expect(wrapper.find('button.c-cart-summary-buttons__cta--proceed')).toHaveProp('disabled', false)
 })
