+++ src/index.js
@@ -100,13 +100,13 @@ export { default as AddressFormHeader } from './components/checkout/address-form
 export { default as AddressFormSummary } from './components/checkout/address-form-summary'
 export { default as CheckoutAddressForm } from './components/checkout/address-form'
 export { default as CheckoutCart } from './components/checkout/checkout-cart'
+export { default as CheckoutCartButtons } from './components/checkout/checkout-cart-buttons'
 export { default as CheckoutCartTotal } from './components/checkout/checkout-cart-total'
 export { default as CheckoutSteps } from './components/checkout/checkout-steps'
 export { default as LineItems } from './components/cart/line-items'
 export { default as MiniPlaceOrder } from './components/checkout/mini-place-order'
 export { default as PaymentIcons } from './components/cart/payment-icons'
 export { default as PaymentMethods } from './components/checkout/payment-methods'
-export { default as PaymentMethodsHeader } from './components/checkout/payment-methods-header'
 export { default as PaymentMethod } from './components/checkout/payment-method'
 export { default as PaymentMethodHeader } from './components/checkout/payment-method-header'
 export { default as PaymentMethodSummary } from './components/checkout/payment-method-summary'
